A lot of lil changes

+++ src/Models/Interfaces/IEntity.cs
@@ -5,7 +5,7 @@ namespace PoeHUD.Models.Interfaces
     public interface IEntity
     {
         string Path { get; }
-        long Id { get; }
+        uint Id { get; }
         bool IsValid { get; }
         bool IsHostile { get; }
         long Address { get; }
